pipeline/pyproject.toml
[project]
name = "ai-tech-digest-pipeline"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "google-genai==1.5.0",
    "feedparser==6.0.12",
    "httpx==0.28.1",
    "crawl4ai==0.8.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.0",
    "pyyaml==6.0.3",
    "thefuzz==0.22.1",
    "chardet==5.2.0",
]

[project.optional-dependencies]
dev = ["pytest==9.0.2", "pytest-asyncio==1.3.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
